Add support for building .mcm files from an image with the whole font
build now accepts either a directory or a .png file as input. Note
that the the margin and number of columns need to match the ones
used in the png. Their default values are the same that we use for
generating the previews, so a preview generated by this tool can be
edited and then turned back into a font.