Add support for inheriting from a default font
Fonts that inherit from the default one don't need to duplicate
their graphics anymore, instead we pull them from the default font.
Also, if a duplicated character is found, a verbose message is printed
to the console to let the user know they can safely delete the file.
Eventually, this will be expanded to allow for more general inheritance.